// src/app/models/realisations.ts
export interface Realisations {
    status?: string;
    message?: string;
    id?: number;
    titre?: string;
    categorie?: string;
    type?: string;
    image?: string;
}

export interface RealisationsTypes {
    status?: string;
    message?: string;
    id?: number;
    name?: string;
}

export interface RealisationsCategories {
    status?: string;
    message?: string;
    id?: number;
    name?: string;
}